pyproject.toml
[project]
name = "event-research-agent"
version = "0.1.0"
description = "AI agent that researches event venues and pushes recommendations to Notion"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "notion-client>=2.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "eval-type-backport>=0.2.0",
]

[project.scripts]
event-research = "event_research.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
